// components/Home/types.ts
export type ArticleImagePropType = {
  imgURL: string;
  alt: string;
};

export type ArticleSummaryBoxPropType = {
  title: string;
  datePublished: string;
};

export type ArticleContainerPropType = {
  id: number | string;
  title: string;
  datePublished: string;
  imgURL: string;
  alt: string;
};